backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyMuPDF==1.23.0
pdf2docx==0.5.8
groq==0.4.2
python-dotenv==1.0.0
pydantic==2.5.0
